Reworked the CI Pipeline to accomodate memote
